requirements.txt
fastapi
uvicorn
sqlalchemy
pydanti